+++ coloraide/color.py
@@ -55,6 +55,7 @@ from .filters.w3c_filter_effects import Sepia, Brightness, Contrast, Saturate, O
 from .filters.cvd import Protan, Deutan, Tritan
 from .interpolate import Interpolator, Interpolate
 from .interpolate.bspline import BSpline
+from .interpolate.bspline_natural import NaturalBSpline
 from .interpolate.linear import Linear
 from .types import Plugin
 from typing import overload, Union, Sequence, Dict, List, Optional, Any, cast, Callable, Tuple, Type, Mapping
@@ -946,6 +947,7 @@ Color.register(
 
         # Interpolation
         Linear(),
-        BSpline()
+        BSpline(),
+        NaturalBSpline()
     ]
 )